// web/lib/api-types.ts
export interface components {
    schemas: {
        CategoriaDto: {
            /** Format: int32 */
            id: number;
            nome: string;
            tipo: components["schemas"]["TipoTransacao"];
            cor: string;
        };
        CategoriaInput: {
            nome: string;
            tipo: components["schemas"]["TipoTransacao"];
            cor: string;
        };
        ContaDto: {
            /** Format: int32 */
            id: number;
            nome: string;
            tipo: components["schemas"]["TipoConta"];
            /** Format: double */
            saldoInicial: number;
            ativa: boolean;
            /** Format: double */
            saldoAtual: number;
        };
        ContaInput: {
            nome: string;
            tipo: components["schemas"]["TipoConta"];
            /** Format: double */
            saldoInicial: number;
            /** @default true */
            ativa: boolean;
        };
        EvolucaoMesDto: {
            /** Format: int32 */
            ano: number;
            /** Format: int32 */
            mes: number;
            /** Format: double */
            receitas: number;
            /** Format: double */
            despesas: number;
            /** Format: double */
            resultado: number;
        };
        GastoCategoriaDto: {
            /** Format: int32 */
            categoriaId: number;
            nome: string;
            cor: string;
            /** Format: double */
            total: number;
            /** Format: double */
            percentual: number;
        };
        HttpValidationProblemDetails: {
            type?: null | string;
            title?: null | string;
            /** Format: int32 */
            status?: null | number;
            detail?: null | string;
            instance?: null | string;
            errors?: {
                [key: string]: string[];
            };
        };
        OrcamentoDto: {
            /** Format: int32 */
            id: null | number;
            /** Format: int32 */
            categoriaId: number;
            categoriaNome: string;
            categoriaCor: string;
            /** Format: int32 */
            ano: number;
            /** Format: int32 */
            mes: number;
            /** Format: double */
            limite: null | number;
            /** Format: double */
            gasto: number;
            /** Format: double */
            percentual: null | number;
        };
        OrcamentoInput: {
            /** Format: int32 */
            categoriaId: number;
            /** Format: int32 */
            ano: number;
            /** Format: int32 */
            mes: number;
            /** Format: double */
            limite: number;
        };
        PaginaOfTransacaoDto: {
            itens: components["schemas"]["TransacaoDto"][];
            /** Format: int32 */
            numeroPagina: number;
            /** Format: int32 */
            tamanhoPagina: number;
            /** Format: int32 */
            totalItens: number;
            /** Format: int32 */
            totalPaginas?: number;
        };
        ResumoDto: {
            /** Format: int32 */
            ano: number;
            /** Format: int32 */
            mes: number;
            /** Format: double */
            saldoTotal: number;
            /** Format: double */
            receitas: number;
            /** Format: double */
            despesas: number;
            /** Format: double */
            resultado: number;
        };
        /** @enum {unknown} */
        TipoConta: "Corrente" | "Poupanca" | "Cartao" | "Carteira" | "Investimento";
        /** @enum {unknown} */
        TipoTransacao: "Receita" | "Despesa";
        TransacaoDto: {
            /** Format: int32 */
            id: number;
            descricao: string;
            /** Format: double */
            valor: number;
            tipo: components["schemas"]["TipoTransacao"];
            /** Format: date */
            data: string;
            /** Format: int32 */
            contaId: number;
            contaNome: string;
            /** Format: int32 */
            categoriaId: number;
            categoriaNome: string;
            categoriaCor: string;
            observacao: null | string;
        };
        TransacaoInput: {
            descricao: string;
            /** Format: double */
            valor: number;
            tipo: components["schemas"]["TipoTransacao"];
            /** Format: date */
            data: string;
            /** Format: int32 */
            contaId: number;
            /** Format: int32 */
            categoriaId: number;
            observacao?: null | string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}